+++ lib/webmock/http_lib_adapters/async_http_client_adapter.rb
@@ -40,8 +40,8 @@ if defined?(Async::HTTP)
         )
           webmock_endpoint = WebMockEndpoint.new(scheme, authority, protocol)
 
-          @network_client = WebMockClient.new(endpoint, protocol, scheme, authority, **options)
-          @webmock_client = WebMockClient.new(webmock_endpoint, protocol, scheme, authority, **options)
+          @network_client = WebMockClient.new(endpoint, **options)
+          @webmock_client = WebMockClient.new(webmock_endpoint, **options)
 
           @scheme = scheme
           @authority = authority
@@ -55,6 +55,7 @@ if defined?(Async::HTTP)
           WebMock::RequestRegistry.instance.requested_signatures.put(request_signature)
           webmock_response = WebMock::StubRegistry.instance.response_for_request(request_signature)
           net_connect_allowed = WebMock.net_connect_allowed?(request_signature.uri)
+          real_request = false
 
           if webmock_response
             webmock_response.raise_error_if_any
@@ -63,6 +64,7 @@ if defined?(Async::HTTP)
             response = @webmock_client.call(request)
           elsif net_connect_allowed
             response = @network_client.call(request)
+            real_request = true
           else
             raise WebMock::NetConnectNotAllowedError.new(request_signature) unless webmock_response
           end
@@ -72,7 +74,7 @@ if defined?(Async::HTTP)
             WebMock::CallbackRegistry.invoke_callbacks(
               {
                 lib: :async_http_client,
-                real_request: net_connect_allowed
+                real_request: real_request
               },
               request_signature,
               webmock_response
@@ -149,7 +151,12 @@ if defined?(Async::HTTP)
         private
 
         def create_connected_sockets
-          Async::IO::Socket.pair(Socket::AF_UNIX, Socket::SOCK_STREAM).tap do |sockets|
+          pair = begin
+            Async::IO::Socket.pair(Socket::AF_UNIX, Socket::SOCK_STREAM)
+          rescue Errno::EAFNOSUPPORT
+            Async::IO::Socket.pair(Socket::AF_INET, Socket::SOCK_STREAM)
+          end
+          pair.tap do |sockets|
             sockets.each do |socket|
               socket.instance_variable_set :@alpn_protocol, nil
               socket.instance_eval do