+++ lib/webmock/http_lib_adapters/http_rb/streamer.rb
@@ -1,8 +1,9 @@
 module HTTP
   class Response
     class Streamer
-      def initialize(str)
+      def initialize(str, encoding: Encoding::BINARY)
         @io = StringIO.new str
+        @encoding = encoding
       end
 
       def readpartial(size = nil, outbuf = nil)
@@ -14,7 +15,8 @@ module HTTP
           end
         end
 
-        @io.read size, outbuf
+        chunk = @io.read size, outbuf
+        chunk.force_encoding(@encoding) if chunk
       end
 
       def close